Refactor of data and folder structure
