+++ src/pyobo/struct/typedef.py
@@ -2,40 +2,16 @@
 
 from __future__ import annotations
 
-import datetime
-from collections import defaultdict
-from collections.abc import Iterable, Mapping, Sequence
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Annotated
+from collections.abc import Sequence
 
 from curies import ReferenceTuple
-from typing_extensions import Self
 
 from . import vocabulary as v
-from .reference import (
-    Reference,
-    Referenced,
-    _reference_list_tag,
-    default_reference,
-    reference_escape,
-)
-from .struct_utils import (
-    AxiomsHint,
-    IntersectionOfHint,
-    PropertiesHint,
-    RelationsHint,
-    Stanza,
-    _chain_tag,
-    _tag_property_targets,
-)
-from .utils import _boolean_tag
+from .reference import Reference
+from .struct import TypeDef
 from ..resources.ro import load_ro
 
-if TYPE_CHECKING:
-    from pyobo.struct.struct import Synonym, SynonymTypeDef
-
 __all__ = [
-    "TypeDef",
     "alternative_term",
     "broad_match",
     "close_match",
@@ -82,278 +58,20 @@ __all__ = [
     "translates_to",
 ]
 
-
-@dataclass
-class TypeDef(Referenced, Stanza):
-    """A type definition in OBO.
-
-    See the subsection of https://owlcollab.github.io/oboformat/doc/GO.format.obo-1_4.html#S.2.2.
-    """
-
-    reference: Annotated[Reference, 1]
-    is_anonymous: Annotated[bool | None, 2] = None
-    # 3 - name is covered by reference
-    namespace: Annotated[str | None, 4] = None
-    # 5 alt_id is part of proerties
-    definition: Annotated[str | None, 6] = None
-    comment: Annotated[str | None, 7] = None
-    subsets: Annotated[list[Reference], 8] = field(default_factory=list)
-    synonyms: Annotated[list[Synonym], 9] = field(default_factory=list)
-    xrefs: Annotated[list[Reference], 10] = field(default_factory=list)
-    _axioms: AxiomsHint = field(default_factory=lambda: defaultdict(list))
-    properties: Annotated[PropertiesHint, 11] = field(default_factory=lambda: defaultdict(list))
-    domain: Annotated[Reference | None, 12, "typedef-only"] = None
-    range: Annotated[Reference | None, 13, "typedef-only"] = None
-    builtin: Annotated[bool | None, 14] = None
-    holds_over_chain: Annotated[list[list[Reference]], 15, "typedef-only"] = field(
-        default_factory=list
-    )
-    is_anti_symmetric: Annotated[bool | None, 16, "typedef-only"] = None
-    is_cyclic: Annotated[bool | None, 17, "typedef-only"] = None
-    is_reflexive: Annotated[bool | None, 18, "typedef-only"] = None
-    is_symmetric: Annotated[bool | None, 19, "typedef-only"] = None
-    is_transitive: Annotated[bool | None, 20, "typedef-only"] = None
-    is_functional: Annotated[bool | None, 21, "typedef-only"] = None
-    is_inverse_functional: Annotated[bool | None, 22, "typedef-only"] = None
-    parents: Annotated[list[Reference], 23] = field(default_factory=list)
-    intersection_of: Annotated[IntersectionOfHint, 24] = field(default_factory=list)
-    union_of: Annotated[list[Reference], 25] = field(default_factory=list)
-    equivalent_to: Annotated[list[Reference], 26] = field(default_factory=list)
-    disjoint_from: Annotated[list[Reference], 27] = field(default_factory=list)
-    # TODO inverse should be inverse_of, cardinality any
-    inverse: Annotated[Reference | None, 28, "typedef-only"] = None
-    # TODO check if there are any examples of this being multiple
-    transitive_over: Annotated[list[Reference], 29, "typedef-only"] = field(default_factory=list)
-    equivalent_to_chain: Annotated[list[list[Reference]], 30, "typedef-only"] = field(
-        default_factory=list
-    )
-    #: From the OBO spec:
-    #:
-    #:   For example: spatially_disconnected_from is disjoint_over part_of, in that two
-    #:   disconnected entities have no parts in common. This can be translated to OWL as:
-    #:   ``disjoint_over(R S), R(A B) ==> (S some A) disjointFrom (S some B)``
-    disjoint_over: Annotated[list[Reference], 31] = field(default_factory=list)
-    relationships: Annotated[RelationsHint, 32] = field(default_factory=lambda: defaultdict(list))
-    is_obsolete: Annotated[bool | None, 33] = None
-    created_by: Annotated[str | None, 34] = None
-    creation_date: Annotated[datetime.datetime | None, 35] = None
-    # TODO expand_assertion_to
-    # TODO expand_expression_to
-    #: Whether this relationship is a metadata tag. Properties that are marked as metadata tags are
-    #: used to record object metadata. Object metadata is additional information about an object
-    #: that is useful to track, but does not impact the definition of the object or how it should
-    #: be treated by a reasoner. Metadata tags might be used to record special term synonyms or
-    #: structured notes about a term, for example.
-    is_metadata_tag: Annotated[bool | None, 40, "typedef-only"] = None
-    is_class_level: Annotated[bool | None, 41] = None
-
-    def __hash__(self) -> int:
-        # have to re-define hash because of the @dataclass
-        return hash((self.__class__, self.prefix, self.identifier))
-
-    def iterate_obo_lines(
-        self,
-        ontology_prefix: str,
-        synonym_typedefs: Mapping[ReferenceTuple, SynonymTypeDef] | None = None,
-        typedefs: Mapping[ReferenceTuple, TypeDef] | None = None,
-    ) -> Iterable[str]:
-        """Iterate over the lines to write in an OBO file.
-
-        :param ontology_prefix:
-            The prefix of the ontology into which the type definition is being written.
-            This is used for compressing builtin identifiers
-        :yield:
-            The lines to write to an OBO file
-
-        `S.3.5.5 <https://owlcollab.github.io/oboformat/doc/GO.format.obo-1_4.html#S.3.5.5>`_
-        of the OBO Flat File Specification v1.4 says tags should appear in the following order:
-
-        1. id
-        2. is_anonymous
-        3. name
-        4. namespace
-        5. alt_id
-        6. def
-        7. comment
-        8. subset
-        9. synonym
-        10. xref
-        11. property_value
-        12. domain
-        13. range
-        14. builtin
-        15. holds_over_chain
-        16. is_anti_symmetric
-        17. is_cyclic
-        18. is_reflexive
-        19. is_symmetric
-        20. is_transitive
-        21. is_functional
-        22. is_inverse_functional
-        23. is_a
-        24. intersection_of
-        25. union_of
-        26. equivalent_to
-        27. disjoint_from
-        28. inverse_of
-        29. transitive_over
-        30. equivalent_to_chain
-        31. disjoint_over
-        32. relationship
-        33. is_obsolete
-        34. created_by
-        35. creation_date
-        36. replaced_by
-        37. consider
-        38. expand_assertion_to
-        39. expand_expression_to
-        40. is_metadata_tag
-        41. is_class_level
-        """
-        if synonym_typedefs is None:
-            synonym_typedefs = {}
-        if typedefs is None:
-            typedefs = {}
-
-        yield "\n[Typedef]"
-        # 1
-        yield f"id: {reference_escape(self.reference, ontology_prefix=ontology_prefix)}"
-        # 2
-        yield from _boolean_tag("is_anonymous", self.is_anonymous)
-        # 3
-        if self.name:
-            yield f"name: {self.name}"
-        # 4
-        if self.namespace:
-            yield f"namespace: {self.namespace}"
-        # 5
-        yield from _reference_list_tag("alt_id", self.alt_ids, ontology_prefix)
-        # 6
-        if self.definition:
-            yield f'def: "{self.definition}"'
-        # 7
-        if self.comment:
-            yield f"comment: {self.comment}"
-        # 8
-        yield from _reference_list_tag("subset", self.subsets, ontology_prefix)
-        # 9
-        for synonym in self.synonyms:
-            yield synonym.to_obo(ontology_prefix=ontology_prefix, synonym_typedefs=synonym_typedefs)
-        # 10
-        yield from self._iterate_xref_obo(ontology_prefix=ontology_prefix)
-        # 11
-        yield from self._iterate_obo_properties(
-            ontology_prefix=ontology_prefix,
-            skip_predicates={
-                term_replaced_by.reference,
-                see_also.reference,
-                alternative_term.reference,
-            },
-            typedefs=typedefs,
-        )
-        # 12
-        if self.domain:
-            yield f"domain: {reference_escape(self.domain, ontology_prefix=ontology_prefix, add_name_comment=True)}"
-        # 13
-        if self.range:
-            yield f"range: {reference_escape(self.range, ontology_prefix=ontology_prefix, add_name_comment=True)}"
-        # 14
-        yield from _boolean_tag("builtin", self.builtin)
-        # 15
-        yield from _chain_tag("holds_over_chain", self.holds_over_chain, ontology_prefix)
-        # 16
-        yield from _boolean_tag("is_anti_symmetric", self.is_anti_symmetric)
-        # 17
-        yield from _boolean_tag("is_cyclic", self.is_cyclic)
-        # 18
-        yield from _boolean_tag("is_reflexive", self.is_reflexive)
-        # 19
-        yield from _boolean_tag("is_symmetric", self.is_symmetric)
-        # 20
-        yield from _boolean_tag("is_transitive", self.is_transitive)
-        # 21
-        yield from _boolean_tag("is_functional", self.is_functional)
-        # 22
-        yield from _boolean_tag("is_inverse_functional", self.is_inverse_functional)
-        # 23
-        yield from _reference_list_tag("is_a", self.parents, ontology_prefix)
-        # 24
-        yield from self._iterate_intersection_of_obo(ontology_prefix=ontology_prefix)
-        # 25
-        yield from _reference_list_tag("union_of", self.union_of, ontology_prefix)
-        # 26
-        yield from _reference_list_tag("equivalent_to", self.equivalent_to, ontology_prefix)
-        # 27
-        yield from _reference_list_tag("disjoint_from", self.disjoint_from, ontology_prefix)
-        # 28
-        if self.inverse:
-            yield f"inverse_of: {reference_escape(self.inverse, ontology_prefix=ontology_prefix, add_name_comment=True)}"
-        # 29
-        yield from _reference_list_tag("transitive_over", self.transitive_over, ontology_prefix)
-        # 30
-        yield from _chain_tag("equivalent_to_chain", self.equivalent_to_chain, ontology_prefix)
-        # 31 disjoint_over, see https://github.com/search?q=%22disjoint_over%3A%22+path%3A*.obo&type=code
-        yield from _reference_list_tag(
-            "disjoint_over", self.disjoint_over, ontology_prefix=ontology_prefix
-        )
-        # 32
-        yield from self._iterate_obo_relations(ontology_prefix=ontology_prefix, typedefs=typedefs)
-        # 33
-        yield from _boolean_tag("is_obsolete", self.is_obsolete)
-        # 34
-        if self.created_by:
-            yield f"created_by: {self.created_by}"
-        # 35
-        if self.creation_date is not None:
-            yield f"creation_date: {self.creation_date.isoformat()}"
-        # 36
-        yield from _tag_property_targets(
-            "replaced_by", self, term_replaced_by, ontology_prefix=ontology_prefix
-        )
-        # 37
-        yield from _tag_property_targets(
-            "consider", self, see_also, ontology_prefix=ontology_prefix
-        )
-        # 38 TODO expand_assertion_to
-        # 39 TODO expand_expression_to
-        # 40
-        yield from _boolean_tag("is_metadata_tag", self.is_metadata_tag)
-        # 41
-        yield from _boolean_tag("is_class_level", self.is_class_level)
-
-    @classmethod
-    def from_triple(cls, prefix: str, identifier: str, name: str | None = None) -> TypeDef:
-        """Create a typedef from a reference."""
-        return cls(reference=Reference(prefix=prefix, identifier=identifier, name=name))
-
-    @classmethod
-    def default(cls, prefix: str, identifier: str, *, name: str | None = None) -> Self:
-        """Construct a default type definition from within the OBO namespace."""
-        return cls(reference=default_reference(prefix, identifier, name=name))
-
-
 RO_PREFIX = "RO"
 BFO_PREFIX = "BFO"
 IAO_PREFIX = "IAO"
 SIO_PREFIX = "SIO"
 
-from_species = TypeDef(
-    reference=Reference(prefix=RO_PREFIX, identifier="0002162", name="in taxon"),
-)
+from_species = TypeDef(reference=v.from_species)
 species_specific = TypeDef(
-    reference=Reference(prefix="debio", identifier="0000007", name="species specific"),
+    reference=v.species_specific,
     definition="X speciesSpecific Y means that Y is a general phenomena, "
     "like a pathway, and X is the version that appears in a species. X should state which"
     "species with RO:0002162 (in taxon)",
 )
-has_left_to_right_reaction = TypeDef(
-    Reference(prefix="debio", identifier="0000007", name="has left-to-right reaction"),
-    is_metadata_tag=True,
-)
-has_right_to_left_reaction = TypeDef(
-    Reference(prefix="debio", identifier="0000008", name="has right-to-left reaction"),
-    is_metadata_tag=True,
-)
+has_left_to_right_reaction = TypeDef(v.has_left_to_right_reaction, is_metadata_tag=True)
+has_right_to_left_reaction = TypeDef(v.has_right_to_left_reaction, is_metadata_tag=True)
 has_bidirectional_reaction = TypeDef(
     Reference(prefix="debio", identifier="0000009", name="has bi-directional reaction"),
     is_metadata_tag=True,
@@ -362,16 +80,8 @@ reaction_enabled_by_molecular_function = TypeDef(
     Reference(prefix="debio", identifier="0000047", name="reaction enabled by molecular function")
 )
 
-part_of = TypeDef(
-    reference=Reference(prefix=BFO_PREFIX, identifier="0000050", name="part of"),
-    comment="Inverse of has_part",
-    inverse=Reference(prefix=BFO_PREFIX, identifier="0000051", name="has part"),
-)
-has_part = TypeDef(
-    reference=Reference(prefix=BFO_PREFIX, identifier="0000051", name="has part"),
-    comment="Inverse of part_of",
-    inverse=Reference(prefix=BFO_PREFIX, identifier="0000050", name="part of"),
-)
+part_of = TypeDef(reference=v.part_of, comment="Inverse of has_part", inverse=v.has_part)
+has_part = TypeDef(reference=v.has_part, comment="Inverse of part_of", inverse=v.part_of)
 occurs_in = TypeDef(
     reference=Reference(prefix=BFO_PREFIX, identifier="BFO:0000066", name="occurs in")
 )
@@ -397,43 +107,19 @@ molecularly_interacts_with = TypeDef(
 located_in = TypeDef(
     reference=Reference(prefix=RO_PREFIX, identifier="0001025", name="located in"),
 )
-exact_match = TypeDef(
-    reference=Reference(prefix="skos", identifier="exactMatch", name="exact match"),
-    is_metadata_tag=True,
-)
-narrow_match = TypeDef(
-    reference=Reference(prefix="skos", identifier="narrowMatch", name="narrow match"),
-    is_metadata_tag=True,
-)
-broad_match = TypeDef(
-    reference=Reference(prefix="skos", identifier="broadMatch", name="broad match"),
-    is_metadata_tag=True,
-)
-close_match = TypeDef(
-    reference=Reference(prefix="skos", identifier="closeMatch", name="close match"),
-    is_metadata_tag=True,
-)
-related_match = TypeDef(
-    reference=Reference(prefix="skos", identifier="relatedMatch", name="related match"),
-    is_metadata_tag=True,
-)
+exact_match = TypeDef(reference=v.exact_match, is_metadata_tag=True)
+narrow_match = TypeDef(reference=v.narrow_match, is_metadata_tag=True)
+broad_match = TypeDef(reference=v.broad_match, is_metadata_tag=True)
+close_match = TypeDef(reference=v.close_match, is_metadata_tag=True)
+related_match = TypeDef(reference=v.related_match, is_metadata_tag=True)
 owl_same_as = TypeDef(
-    reference=Reference(prefix="owl", identifier="sameAs", name="same as"),
+    reference=v.owl_same_as,
 )
-equivalent_class = TypeDef(
-    reference=v.equivalent_class,
-)
-equivalent_property = TypeDef(
-    reference=Reference(prefix="owl", identifier="equivalentProperty", name="equivalent property"),
-)
+equivalent_class = TypeDef(reference=v.equivalent_class)
+equivalent_property = TypeDef(reference=v.equivalent_property)
 
-is_a = TypeDef(
-    reference=Reference(prefix="rdfs", identifier="subClassOf", name="subclass of"),
-)
-see_also = TypeDef(
-    reference=v.see_also,
-    is_metadata_tag=True,
-)
+is_a = TypeDef(reference=v.is_a)
+see_also = TypeDef(reference=v.see_also, is_metadata_tag=True)
 comment = TypeDef(reference=v.comment, is_metadata_tag=True)
 has_member = TypeDef(
     reference=Reference(prefix=RO_PREFIX, identifier="0002351", name="has member"),
@@ -449,12 +135,7 @@ superclass_of = TypeDef(
 )
 
 develops_from = TypeDef.from_triple(prefix=RO_PREFIX, identifier="0002202", name="develops from")
-orthologous = TypeDef(
-    reference=Reference(
-        prefix=RO_PREFIX, identifier="HOM0000017", name="in orthology relationship with"
-    ),
-    is_symmetric=True,
-)
+orthologous = TypeDef(reference=v.orthologous, is_symmetric=True)
 
 has_role = TypeDef(
     reference=Reference(prefix=RO_PREFIX, identifier="0000087", name="has role"),
@@ -505,18 +186,12 @@ has_salt = TypeDef(
     reference=Reference(prefix="debio", identifier="0000006", name="has salt"),
 )
 
-term_replaced_by = TypeDef(
-    reference=v.term_replaced_by,
-    is_metadata_tag=True,
-)
+term_replaced_by = TypeDef(reference=v.term_replaced_by, is_metadata_tag=True)
 example_of_usage = TypeDef(
     reference=Reference(prefix=IAO_PREFIX, identifier="0000112", name="example of usage"),
     is_metadata_tag=True,
 )
-alternative_term = TypeDef(
-    reference=v.alternative_term,
-    is_metadata_tag=True,
-)
+alternative_term = TypeDef(reference=v.alternative_term, is_metadata_tag=True)
 has_ontology_root_term = TypeDef(reference=v.has_ontology_root_term, is_metadata_tag=True)
 definition_source = TypeDef(
     reference=Reference(prefix=IAO_PREFIX, identifier="0000119", name="definition source"),
@@ -602,10 +277,7 @@ mapping_has_confidence = TypeDef(
     is_metadata_tag=True,
     range=Reference(prefix="xsd", identifier="float"),
 )
-has_contributor = TypeDef(
-    reference=v.has_contributor,
-    is_metadata_tag=True,
-)
+has_contributor = TypeDef(reference=v.has_contributor, is_metadata_tag=True)
 
 has_start_date = TypeDef(
     reference=Reference(prefix="dcat", identifier="startDate", name="has start date"),
@@ -616,18 +288,9 @@ has_end_date = TypeDef(
     is_metadata_tag=True,
 )
 
-has_title = TypeDef(
-    reference=Reference(prefix="dcterms", identifier="title", name="title"),
-    is_metadata_tag=True,
-)
-has_license = TypeDef(
-    reference=Reference(prefix="dcterms", identifier="license", name="license"),
-    is_metadata_tag=True,
-)
-has_description = TypeDef(
-    reference=Reference(prefix="dcterms", identifier="description", name="description"),
-    is_metadata_tag=True,
-)
+has_title = TypeDef(reference=v.has_title, is_metadata_tag=True)
+has_license = TypeDef(reference=v.has_license, is_metadata_tag=True)
+has_description = TypeDef(reference=v.has_description, is_metadata_tag=True)
 
 default_typedefs: dict[ReferenceTuple, TypeDef] = {
     v.pair: v for v in locals().values() if isinstance(v, TypeDef)
